Add files via upload
Code is used to create a graph to visualize excitation scans over multiple wavelengths (x-axis) and the mean intensity values normalized to power measurements (y-axis). Input will be CSV file of segmented transfected, untransfected, and cytoplasm of cells transfected with different fluorophores. Output will be a .png file with the visualization along with the new CSV file with all the data graphed.